+++ com.unity.render-pipelines.universal/Runtime/Data/UniversalRenderPipelineAsset.cs
@@ -19,30 +19,97 @@ namespace UnityEngine.Rendering.Universal
         SoftShadows,
     }
 
+    /// <summary>
+    /// This controls the size of the shadow map texture.
+    /// </summary>
     public enum ShadowResolution
     {
+        /// <summary>
+        /// Use this for 256x256 shadow resolution.
+        /// </summary>
         _256 = 256,
+
+        /// <summary>
+        /// Use this for 512x512 shadow resolution.
+        /// </summary>
         _512 = 512,
+
+        /// <summary>
+        /// Use this for 1024x1024 shadow resolution.
+        /// </summary>
         _1024 = 1024,
+
+        /// <summary>
+        /// Use this for 2048x2048 shadow resolution.
+        /// </summary>
         _2048 = 2048,
+
+        /// <summary>
+        /// Use this for 4096x4096 shadow resolution.
+        /// </summary>
         _4096 = 4096
     }
 
+    /// <summary>
+    /// This controls the size of the Light Cookie atlas texture for additional lights (point, spot).
+    /// </summary>
     public enum LightCookieResolution
     {
+        /// <summary>
+        /// Use this for 256x256 Light Cookie resolution.
+        /// </summary>
         _256 = 256,
+
+        /// <summary>
+        /// Use this for 512x512 Light Cookie resolution.
+        /// </summary>
         _512 = 512,
+
+        /// <summary>
+        /// Use this for 1024x1024 Light Cookie resolution.
+        /// </summary>
         _1024 = 1024,
+
+        /// <summary>
+        /// Use this for 2048x2048 Light Cookie resolution.
+        /// </summary>
         _2048 = 2048,
+
+        /// <summary>
+        /// Use this for 4096x4096 Light Cookie resolution.
+        /// </summary>
         _4096 = 4096
     }
 
+    /// <summary>
+    /// Options for selecting the format for the Light Cookie atlas texture for additional lights (point, spot).
+    /// Low precision saves memory and bandwidth.
+    /// </summary>
     public enum LightCookieFormat
     {
+        /// <summary>
+        /// Use this to select Grayscale format with low precision.
+        /// </summary>
         GrayscaleLow,
+
+        /// <summary>
+        /// Use this to select Grayscale format with high precision.
+        /// </summary>
         GrayscaleHigh,
+
+        /// <summary>
+        /// Use this to select Color format with low precision.
+        /// </summary>
         ColorLow,
+
+        /// <summary>
+        /// Use this to select Color format with high precision.
+        /// </summary>
         ColorHigh,
+
+        /// <summary>
+        /// Use this to select High Dynamic Range format.
+        /// </summary>
         ColorHDR,
     }
 
@@ -58,19 +125,56 @@ namespace UnityEngine.Rendering.Universal
         _64Bits,
     }
 
+    /// <summary>
+    /// Options for setting MSAA Quality.
+    /// This defines how many samples URP computes per pixel for evaluating the effect.
+    /// </summary>
     public enum MsaaQuality
     {
+        /// <summary>
+        /// Disables MSAA.
+        /// </summary>
         Disabled = 1,
+
+        /// <summary>
+        /// Use this for 2 samples per pixel.
+        /// </summary>
         _2x = 2,
+
+        /// <summary>
+        /// Use this for 4 samples per pixel.
+        /// </summary>
         _4x = 4,
+
+        /// <summary>
+        /// Use this for 8 samples per pixel.
+        /// </summary>
         _8x = 8
     }
 
+    /// <summary>
+    /// Options for selecting downsampling.
+    /// </summary>
     public enum Downsampling
     {
+        /// <summary>
+        /// Use this to disable downsampling.
+        /// </summary>
         None,
+
+        /// <summary>
+        /// Use this to produce a half-resolution image with bilinear filtering.
+        /// </summary>
         _2xBilinear,
+
+        /// <summary>
+        /// Use this to produce a quarter-resolution image with box filtering. This produces a softly blurred copy.
+        /// </summary>
         _4xBox,
+
+        /// <summary>
+        /// Use this to produce a quarter-resolution image with bi-linear filtering.
+        /// </summary>
         _4xBilinear
     }
 
@@ -85,10 +189,24 @@ namespace UnityEngine.Rendering.Universal
         Decal
     }
 
+    /// <summary>
+    /// Options for light rendering mode.
+    /// </summary>
     public enum LightRenderingMode
     {
+        /// <summary>
+        /// Use this to disable lighting.
+        /// </summary>
         Disabled = 0,
+
+        /// <summary>
+        /// Use this to select lighting to be calculated per vertex.
+        /// </summary>
         PerVertex = 2,
+
+        /// <summary>
+        /// Use this to select lighting to be calculated per pixel.
+        /// </summary>
         PerPixel = 1,
     }
 
@@ -99,18 +217,43 @@ namespace UnityEngine.Rendering.Universal
         Profiling,
     }
 
+    /// <summary>
+    /// Options to select the type of Renderer to use.
+    /// </summary>
     public enum RendererType
     {
+        /// <summary>
+        /// Use this for Custom Renderer.
+        /// </summary>
         Custom,
+
+        /// <summary>
+        /// Use this for Universal Renderer.
+        /// </summary>
         UniversalRenderer,
+
+        /// <summary>
+        /// Use this for 2D Renderer.
+        /// </summary>
         _2DRenderer,
+
         [Obsolete("ForwardRenderer has been renamed (UnityUpgradable) -> UniversalRenderer", true)]
         ForwardRenderer = UniversalRenderer,
     }
 
+    /// <summary>
+    /// Options for selecting Color Grading modes, Low Dynamic Range (LDR) or High Dynamic Range (HDR)
+    /// </summary>
     public enum ColorGradingMode
     {
+        /// <summary>
+        /// This mode follows a more classic workflow. Unity applies a limited range of color grading after tonemapping.
+        /// </summary>
         LowDynamicRange,
+
+        /// <summary>
+        /// This mode works best for high precision grading similar to movie production workflows. Unity applies color grading before tonemapping.
+        /// </summary>
         HighDynamicRange
     }
 
@@ -132,10 +275,21 @@ namespace UnityEngine.Rendering.Universal
     /// </summary>
     public enum VolumeFrameworkUpdateMode
     {
+        /// <summary>
+        /// Use this to have the Volume Framework update every frame.
+        /// </summary>
         [InspectorName("Every Frame")]
         EveryFrame = 0,
+
+        /// <summary>
+        /// Use this to disable Volume Framework updates or to update it manually via scripting.
+        /// </summary>
         [InspectorName("Via Scripting")]
         ViaScripting = 1,
+
+        /// <summary>
+        /// Use this to choose the setting set on the pipeline asset.
+        /// </summary>
         [InspectorName("Use Pipeline Settings")]
         UsePipelineSettings = 2,
     }
@@ -145,12 +299,25 @@ namespace UnityEngine.Rendering.Universal
     /// </summary>
     public enum UpscalingFilterSelection
     {
+        /// <summary>
+        /// Indicates that URP will select an appropriate upscaling filter automatically.
+        /// </summary>
         [InspectorName("Automatic")]
         Auto,
+
+        /// <summary>
+        /// Indicates that Bilinear filtering will be used when performing upscaling.
+        /// </summary>
         [InspectorName("Bilinear")]
         Linear,
+
+        /// <summary>
+        /// Indicates that Nearest-Neighbour filtering will be used when performing upscaling.
+        /// </summary>
         [InspectorName("Nearest-Neighbor")]
-        Point
+        Point,
+        [InspectorName("FidelityFX Super Resolution 1.0")]
+        FSR
     }
 
     [ExcludeFromPreset]
@@ -185,6 +352,8 @@ namespace UnityEngine.Rendering.Universal
         [SerializeField] MsaaQuality m_MSAA = MsaaQuality.Disabled;
         [SerializeField] float m_RenderScale = 1.0f;
         [SerializeField] UpscalingFilterSelection m_UpscalingFilter = UpscalingFilterSelection.Auto;
+        [SerializeField] bool m_FsrOverrideSharpness = false;
+        [SerializeField] float m_FsrSharpness = FSRUtils.kDefaultSharpnessLinear;
         // TODO: Shader Quality Tiers
 
         // Main directional light Settings
@@ -251,15 +420,15 @@ namespace UnityEngine.Rendering.Universal
         // Note: A lut size of 16^3 is barely usable with the HDR grading mode. 32 should be the
         // minimum, the lut being encoded in log. Lower sizes would work better with an additional
         // 1D shaper lut but for now we'll keep it simple.
-        public const int k_MinLutSize = 16;
-        public const int k_MaxLutSize = 65;
+        internal const int k_MinLutSize = 16;
+        internal const int k_MaxLutSize = 65;
 
         internal const int k_ShadowCascadeMinCount = 1;
         internal const int k_ShadowCascadeMaxCount = 4;
 
-        public static readonly int AdditionalLightsDefaultShadowResolutionTierLow = 256;
-        public static readonly int AdditionalLightsDefaultShadowResolutionTierMedium = 512;
-        public static readonly int AdditionalLightsDefaultShadowResolutionTierHigh = 1024;
+        internal static readonly int AdditionalLightsDefaultShadowResolutionTierLow = 256;
+        internal static readonly int AdditionalLightsDefaultShadowResolutionTierMedium = 512;
+        internal static readonly int AdditionalLightsDefaultShadowResolutionTierHigh = 1024;
 
 #if UNITY_EDITOR
         [NonSerialized]
@@ -421,6 +590,7 @@ namespace UnityEngine.Rendering.Universal
             }
         }
 
+        /// <inheritdoc/>
         protected override void OnValidate()
         {
             DestroyRenderers();
@@ -430,6 +600,7 @@ namespace UnityEngine.Rendering.Universal
             base.OnValidate();
         }
 
+        /// <inheritdoc/>
         protected override void OnDisable()
         {
             DestroyRenderers();
@@ -710,6 +881,28 @@ namespace UnityEngine.Rendering.Universal
             set { m_UpscalingFilter = value; }
         }
 
+        /// <summary>
+        /// If this property is set to true, the value from the fsrSharpness property will control the intensity of the
+        /// sharpening filter associated with FidelityFX Super Resolution.
+        /// </summary>
+        public bool fsrOverrideSharpness
+        {
+            get { return m_FsrOverrideSharpness; }
+            set { m_FsrOverrideSharpness = value; }
+        }
+
+        /// <summary>
+        /// Controls the intensity of the sharpening filter associated with FidelityFX Super Resolution.
+        /// A value of 1.0 produces maximum sharpness while a value of 0.0 disables the sharpening filter entirely.
+        ///
+        /// Note: This value only has an effect when the fsrOverrideSharpness property is set to true.
+        /// </summary>
+        public float fsrSharpness
+        {
+            get { return m_FsrSharpness; }
+            set { m_FsrSharpness = value; }
+        }
+
         public LightRenderingMode mainLightRenderingMode
         {
             get { return m_MainLightRenderingMode; }
@@ -932,12 +1125,18 @@ namespace UnityEngine.Rendering.Universal
             set { m_UseSRPBatcher = value; }
         }
 
+        /// <summary>
+        /// Returns the selected ColorGradingMode in the URP Asset.
+        /// </summary>
         public ColorGradingMode colorGradingMode
         {
             get { return m_ColorGradingMode; }
             set { m_ColorGradingMode = value; }
         }
 
+        /// <summary>
+        /// Returns the selected LUT size for Color Grading in the URP Asset.
+        /// </summary>
         public int colorGradingLutSize
         {
             get { return m_ColorGradingLutSize; }