+++ com.unity.render-pipelines.universal/Runtime/ForwardLights.cs
@@ -88,6 +88,9 @@ namespace UnityEngine.Rendering.Universal.Internal
             }
         }
 
+        /// <summary>
+        /// Creates a new <c>ForwardLights</c> instance.
+        /// </summary>
         public ForwardLights() : this(InitParams.Create()) { }
 
         internal ForwardLights(InitParams initParams)
@@ -323,6 +326,11 @@ namespace UnityEngine.Rendering.Universal.Internal
             }
         }
 
+        /// <summary>
+        /// Sets up the keywords and data for forward lighting.
+        /// </summary>
+        /// <param name="context"></param>
+        /// <param name="renderingData"></param>
         public void Setup(ScriptableRenderContext context, ref RenderingData renderingData)
         {
             int additionalLightsCount = renderingData.lightData.additionalLightsCount;