+++ com.unity.shadergraph/Editor/Data/Graphs/GraphData.cs
@@ -1088,6 +1088,11 @@ namespace UnityEditor.ShaderGraph
             return newEdge;
         }
 
+        internal void UnnotifyAddedEdge(IEdge edge)
+        {
+            m_AddedEdges.Remove(edge);
+        }
+
         public void RemoveEdge(IEdge e)
         {
             RemoveEdgeNoValidate(e);
@@ -1874,6 +1879,16 @@ namespace UnityEditor.ShaderGraph
                     m_ParentGroupChanges.Remove(groupChange);
                 }
             }
+
+            var existingDefaultCategory = categories.FirstOrDefault();
+            if (existingDefaultCategory?.childCount == 0 && categories.Count() == 1 && (properties.Count() != 0 || keywords.Count() != 0 || dropdowns.Count() != 0))
+            {
+                // Have a graph with category data in invalid state
+                // there is only one category, the default category, and all shader inputs should belong to it
+                // Clear category data as it will get reconstructed in the BlackboardController constructor
+                m_CategoryData.Clear();
+            }
+
 
             ValidateCustomBlockLimit();
             ValidateContextBlocks();